requirements.txt
certifi==2019.3.9
chardet==3.0.4
clearbit==0.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
flake8==3.7.7
idna==2.8
isort==4.3.17
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
pyhunter==0.3
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
